refactor: Eliminate lib -> app dependency
