+++ src/JsonSettingsIO.h
@@ -3,7 +3,6 @@
 class CrossPointSettings;
 class CrossPointState;
 class WifiCredentialStore;
-class KOReaderCredentialStore;
 class RecentBooksStore;
 class OpdsServerStore;
 
@@ -20,10 +19,6 @@ bool loadState(CrossPointState& s, const char* json);
 // WifiCredentialStore
 bool saveWifi(const WifiCredentialStore& store, const char* path);
 bool loadWifi(WifiCredentialStore& store, const char* json, bool* needsResave = nullptr);
-
-// KOReaderCredentialStore
-bool saveKOReader(const KOReaderCredentialStore& store, const char* path);
-bool loadKOReader(KOReaderCredentialStore& store, const char* json, bool* needsResave = nullptr);
 
 // RecentBooksStore
 bool saveRecentBooks(const RecentBooksStore& store, const char* path);
